requirements.txt
# 핵심
ib_insync>=0.9.86
lightgbm>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
scikit-learn>=1.3.0

# 데이터 (초기 다운로드)
databento>=0.30.0

# 선택 (경제지표 웹 캘린더)
# aiohttp>=3.9.0
# beautifulsoup4>=4.12.0
